Отказ перетусовки классов между пакетами
