+++ src/main/java/com/github/_1c_syntax/mdclasses/mdo/MDObjectBase.java
@@ -23,7 +23,7 @@ package com.github._1c_syntax.mdclasses.mdo;
 
 import com.github._1c_syntax.mdclasses.mdo.wrapper.DesignerMDO;
 import com.github._1c_syntax.mdclasses.mdo.wrapper.DesignerSynonym;
-import com.github._1c_syntax.mdclasses.mdo.additional.MDOReference;
+import com.github._1c_syntax.mdclasses.metadata.additional.MDOReference;
 import com.github._1c_syntax.mdclasses.metadata.additional.MDOType;
 import com.github._1c_syntax.mdclasses.metadata.additional.ObjectBelonging;
 import com.thoughtworks.xstream.annotations.XStreamAsAttribute;
