+++ src/main/java/com/github/_1c_syntax/mdclasses/metadata/additional/DataPath.java
@@ -19,16 +19,21 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with MDClasses.
  */
-package com.github._1c_syntax.mdclasses.mdo.additional;
+package com.github._1c_syntax.mdclasses.metadata.additional;
 
-import com.thoughtworks.xstream.annotations.XStreamImplicit;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
-import java.util.Collections;
-import java.util.List;
-
+/**
+ * Модель пути к данным. Используется в элементах формы
+ */
 @Data
-public class ValueType {
-  @XStreamImplicit
-  List<String> types = Collections.emptyList();
+@NoArgsConstructor
+public class DataPath {
+  public static final DataPath EMPTY = new DataPath();
+  private String segment = "";
+
+  public DataPath(String segment) {
+    this.setSegment(segment);
+  }
 }
