+++ src/main/java/com/github/_1c_syntax/mdclasses/unmarshal/XStreamFactory.java
@@ -71,10 +71,10 @@ import com.github._1c_syntax.mdclasses.mdo.Resource;
 import com.github._1c_syntax.mdclasses.mdo.Right;
 import com.github._1c_syntax.mdclasses.mdo.Template;
 import com.github._1c_syntax.mdclasses.mdo.WEBServiceOperation;
-import com.github._1c_syntax.mdclasses.mdo.additional.DataPath;
-import com.github._1c_syntax.mdclasses.mdo.additional.MDOModule;
-import com.github._1c_syntax.mdclasses.mdo.additional.MDOReference;
-import com.github._1c_syntax.mdclasses.mdo.additional.ValueType;
+import com.github._1c_syntax.mdclasses.metadata.additional.DataPath;
+import com.github._1c_syntax.mdclasses.metadata.additional.MDOModule;
+import com.github._1c_syntax.mdclasses.metadata.additional.MDOReference;
+import com.github._1c_syntax.mdclasses.metadata.additional.ValueType;
 import com.github._1c_syntax.mdclasses.mdo.form.FormAttribute;
 import com.github._1c_syntax.mdclasses.mdo.form.FormCommand;
 import com.github._1c_syntax.mdclasses.mdo.form.FormData;
